+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 # Change Log
 
 ## [v1.0.0]
-- Hyperbola ()
+
+Breaking changes
+- Hyperbola: methods for creation are changes, now you can use segment.
+- Ellipse: methods for creation are changes, now you can use segment.
+
+News
+- Hyperbola: you can have some characteristics of the hyperbola.
+- Add some test and bug fixing.
 
 ## [v0.0.1]
 - First release.